eval statement refactor
